Rename `NestedDic` to `DotDict`
